+++ packages/pigeon/platform_tests/test_plugin/macos/Classes/TestPlugin.swift
@@ -23,90 +23,90 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
 
   // MARK: HostIntegrationCoreApi implementation
 
-    func noop() {
+  public func noop() {
   }
 
-  func echoAllTypes(everything: AllTypes) -> AllTypes {
+  public func echoAllTypes(everything: AllTypes) -> AllTypes {
     return everything
   }
 
-  func echoAllNullableTypes(everything: AllNullableTypes?) -> AllNullableTypes? {
+  public func echoAllNullableTypes(everything: AllNullableTypes?) -> AllNullableTypes? {
     return everything
   }
 
-  func throwError() {
+  public func throwError() {
     // TODO(stuartmorgan): Implement this. See
     // https://github.com/flutter/flutter/issues/112483
   }
 
-  func echoInt(anInt: Int32) -> Int32 {
+  public func echoInt(anInt: Int32) -> Int32 {
     return anInt
   }
 
-  func echoDouble(aDouble: Double) -> Double {
+  public func echoDouble(aDouble: Double) -> Double {
     return aDouble
   }
 
-  func echoBool(aBool: Bool) -> Bool {
+  public func echoBool(aBool: Bool) -> Bool {
     return aBool
   }
 
-  func echoString(aString: String) -> String {
+  public func echoString(aString: String) -> String {
     return aString
   }
 
-  func echoUint8List(aUint8List: FlutterStandardTypedData) -> FlutterStandardTypedData {
+  public func echoUint8List(aUint8List: FlutterStandardTypedData) -> FlutterStandardTypedData {
     return aUint8List
   }
 
-  func extractNestedNullableString(wrapper: AllNullableTypesWrapper) -> String? {
+  public func extractNestedNullableString(wrapper: AllNullableTypesWrapper) -> String? {
     return wrapper.values.aNullableString;
   }
 
-  func createNestedNullableString(nullableString: String?) -> AllNullableTypesWrapper {
+  public func createNestedNullableString(nullableString: String?) -> AllNullableTypesWrapper {
     return AllNullableTypesWrapper(values: AllNullableTypes(aNullableString: nullableString))
   }
 
-  func sendMultipleNullableTypes(aNullableBool: Bool?, aNullableInt: Int32?, aNullableString: String?) -> AllNullableTypes {
+  public func sendMultipleNullableTypes(aNullableBool: Bool?, aNullableInt: Int32?, aNullableString: String?) -> AllNullableTypes {
     let someThings = AllNullableTypes(aNullableBool: aNullableBool, aNullableInt: aNullableInt, aNullableString: aNullableString)
     return someThings
   }
 
-  func echoNullableInt(aNullableInt: Int32?) -> Int32? {
+  public func echoNullableInt(aNullableInt: Int32?) -> Int32? {
     return aNullableInt
   }
 
-  func echoNullableDouble(aNullableDouble: Double?) -> Double? {
+  public func echoNullableDouble(aNullableDouble: Double?) -> Double? {
     return aNullableDouble
   }
 
-  func echoNullableBool(aNullableBool: Bool?) -> Bool? {
+  public func echoNullableBool(aNullableBool: Bool?) -> Bool? {
     return aNullableBool
   }
 
-  func echoNullableString(aNullableString: String?) -> String? {
+  public func echoNullableString(aNullableString: String?) -> String? {
     return aNullableString
   }
 
-  func echoNullableUint8List(aNullableUint8List: FlutterStandardTypedData?) -> FlutterStandardTypedData? {
+  public func echoNullableUint8List(aNullableUint8List: FlutterStandardTypedData?) -> FlutterStandardTypedData? {
     return aNullableUint8List
   }
 
-  func noopAsync(completion: @escaping () -> Void) {
+  public func noopAsync(completion: @escaping () -> Void) {
     completion()
   }
 
-  func echoAsyncString(aString: String, completion: @escaping (String) -> Void) {
+  public func echoAsyncString(aString: String, completion: @escaping (String) -> Void) {
     completion(aString)
   }
 
-  func callFlutterNoop(completion: @escaping () -> Void) {
+  public func callFlutterNoop(completion: @escaping () -> Void) {
     flutterAPI.noop() {
       completion()
     }
   }
 
-  func callFlutterEchoString(aString: String, completion: @escaping (String) -> Void) {
+  public func callFlutterEchoString(aString: String, completion: @escaping (String) -> Void) {
     flutterAPI.echoString(aString: aString) { flutterString in
       completion(flutterString)
     }