+++ packages/pigeon/CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ## 4.2.9
 
+* Added public keyword to relevant structs, protocol and classes
 * [kotlin] Fixes a bug with some methods that return `void`.
 
 ## 4.2.8